Refs #114. Moved testcases to their own module
