+++ pydocx/test/__init__.py
@@ -1,0 +1,9 @@
+from pydocx.test.testcases import (
+    DocumentGeneratorTestCase,
+    _TranslationTestCase,
+)
+
+__all__ = [
+    'DocumentGeneratorTestCase',
+    '_TranslationTestCase',
+]
