+++ pydocx/test/utils.py
@@ -7,9 +7,7 @@ from __future__ import (
 import posixpath
 import re
 import os.path
-from contextlib import contextmanager
 from xml.dom import minidom
-from unittest import TestCase
 
 try:
     from cString import StringIO
@@ -21,7 +19,6 @@ from pydocx.managers.styles import StylesManager
 from pydocx.packaging import PackageRelationship, ZipPackagePart
 from pydocx.parsers.Docx2Html import Docx2Html
 from pydocx.util.xml import parse_xml_from_string
-from pydocx.util.zip import create_zip_archive
 from pydocx.wordml import (
     FootnotesPart,
     MainDocumentPart,
@@ -31,42 +28,6 @@ from pydocx.wordml import (
 )
 
 from pydocx.test.document_builder import DocxBuilder as DXB
-
-STYLE = (
-    '<style>'
-    '.pydocx-caps {text-transform:uppercase}'
-    '.pydocx-center {text-align:center}'
-    '.pydocx-comment {color:blue}'
-    '.pydocx-delete {color:red;text-decoration:line-through}'
-    '.pydocx-hidden {visibility:hidden}'
-    '.pydocx-insert {color:green}'
-    '.pydocx-left {text-align:left}'
-    '.pydocx-right {text-align:right}'
-    '.pydocx-small-caps {font-variant:small-caps}'
-    '.pydocx-strike {text-decoration:line-through}'
-    '.pydocx-tab {display:inline-block;width:4em}'
-    '.pydocx-underline {text-decoration:underline}'
-    'body {margin:0px auto;width:51.00em}'
-    '</style>'
-)
-
-BASE_HTML = '''
-<html>
-    <head>
-    <meta charset="utf-8" />
-    %s
-    </head>
-    <body>%%s</body>
-</html>
-''' % STYLE
-
-
-BASE_HTML_NO_STYLE = '''
-<html>
-    <head><meta charset="utf-8" /></head>
-    <body>%s</body>
-</html>
-'''
 
 
 def prettify(xml_string):
@@ -303,25 +264,6 @@ class WordprocessingDocumentFactory(object):
         return '[Content_Types].xml', content_types
 
 
-class DocumentGeneratorTestCase(TestCase):
-    '''
-    A test case class that can be inherited to compare xml fragments with their
-    resulting HTML output.
-
-    Each test case needs to call `assert_document_generates_html`
-    '''
-
-    def assert_document_generates_html(self, document, expected_html):
-        zip_buf = create_zip_archive(document.to_zip_dict())
-        parser = Docx2HtmlNoStyle(zip_buf)
-        actual = parser.parsed
-        expected = BASE_HTML_NO_STYLE % expected_html
-        if not html_is_equal(actual, expected):
-            actual = prettify(actual)
-            message = 'The expected HTML did not match the actual HTML:'
-            raise AssertionError(message + '\n' + actual)
-
-
 class XMLDocx2Html(Docx2Html):
     """
     Create the object without passing in a path to the document, set them
@@ -404,68 +346,3 @@ class XMLDocx2Html(Docx2Html):
             return self.numbering_dict[num_id][ilvl]
         except KeyError:
             return 'decimal'
-
-
-DEFAULT_NUMBERING_DICT = {
-    '1': {
-        '0': 'decimal',
-        '1': 'decimal',
-    },
-    '2': {
-        '0': 'lowerLetter',
-        '1': 'lowerLetter',
-    },
-}
-
-
-class _TranslationTestCase(TestCase):
-    expected_output = None
-    relationships = None
-    numbering_dict = DEFAULT_NUMBERING_DICT
-    run_expected_output = True
-    parser = XMLDocx2Html
-    use_base_html = True
-    convert_root_level_upper_roman = False
-    styles_xml = None
-
-    def get_xml(self):
-        raise NotImplementedError()
-
-    @contextmanager
-    def toggle_run_expected_output(self):
-        self.run_expected_output = not self.run_expected_output
-        yield
-        self.run_expected_output = not self.run_expected_output
-
-    def test_expected_output(self):
-        if self.expected_output is None:
-            raise NotImplementedError('expected_output is not defined')
-        if not self.run_expected_output:
-            return
-
-        # Create the xml
-        tree = self.get_xml()
-
-        # Verify the final output.
-        parser = self.parser
-
-        html = parser(
-            convert_root_level_upper_roman=self.convert_root_level_upper_roman,
-            document_xml=tree,
-            relationships=self.relationships,
-            numbering_dict=self.numbering_dict,
-            styles_xml=self.styles_xml,
-        ).parsed
-
-        if self.use_base_html:
-            assert_html_equal(
-                html,
-                BASE_HTML % self.expected_output,
-                filename=self.__class__.__name__,
-            )
-        else:
-            assert_html_equal(
-                html,
-                self.expected_output,
-                filename=self.__class__.__name__,
-            )
